feat: Restart command for restarting only one container

+++ plugins/docker/oper-restart.sh
@@ -2,10 +2,21 @@
 : "${taito_cli_path:?}"
 : "${taito_plugin_path:?}"
 
-"${taito_cli_path}/util/execute-on-host-fg.sh" \
-  "docker-compose down" && \
-"${taito_cli_path}/util/execute-on-host-fg.sh" \
-  "${taito_plugin_path}/util/start.sh" "${@}" && \
+pod="${1}"
+
+if [[ -n "${pod}" ]]; then
+  # Restart only the container given as argument
+  # shellcheck disable=SC1090
+  . "${taito_plugin_path}/util/determine-pod.sh" "${@}" && \
+  "${taito_cli_path}/util/execute-on-host-fg.sh" \
+    "docker-compose restart ${pod:?}"
+else
+  # Restart all containers
+  "${taito_cli_path}/util/execute-on-host-fg.sh" \
+    "docker-compose down" && \
+  "${taito_cli_path}/util/execute-on-host-fg.sh" \
+    "${taito_plugin_path}/util/start.sh" "${@}"
+fi && \
 
 # Call next command on command chain
 "${taito_cli_path}/util/call-next.sh" "${@}"
